refactor!: move legacy rulesets into lib/legacy/
