+++ lib/legacy/react.mjs
